requirements.txt
asgiref==3.11.1
Django>=5.0,<5.2
pillow==12.1.1
sqlparse==0.5.5
psycopg2-binary
python-dotenv
django-jazzmin
python-telegram-bot
